LWM2M: merge with master
