+++ common/transport/lwm2m/src/main/java/org/thingsboard/server/transport/lwm2m/server/ota/firmware/UpdateStateFw.java
@@ -13,36 +13,44 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.thingsboard.server.transport.lwm2m.server;
+package org.thingsboard.server.transport.lwm2m.server.ota.firmware;
 
-public enum LwM2MSoftwareUpdateStrategy {
-    BINARY(1, "ObjectId 9, Binary"),
-    TEMP_URL(2, "ObjectId 9, URI");
+/**
+ * /** State R
+ * 0: Idle (before downloading or after successful updating)
+ * 1: Downloading (The data sequence is on the way)
+ * 2: Downloaded
+ * 3: Updating
+ */
+public enum UpdateStateFw {
+    IDLE(0, "Idle"),
+    DOWNLOADING(1, "Downloading"),
+    DOWNLOADED(2, "Downloaded"),
+    UPDATING(3, "Updating");
 
     public int code;
     public String type;
 
-    LwM2MSoftwareUpdateStrategy(int code, String type) {
+    UpdateStateFw(int code, String type) {
         this.code = code;
         this.type = type;
     }
 
-    public static LwM2MSoftwareUpdateStrategy fromStrategySwByType(String type) {
-        for (LwM2MSoftwareUpdateStrategy to : LwM2MSoftwareUpdateStrategy.values()) {
+    public static UpdateStateFw fromStateFwByType(String type) {
+        for (UpdateStateFw to : UpdateStateFw.values()) {
             if (to.type.equals(type)) {
                 return to;
             }
         }
-        throw new IllegalArgumentException(String.format("Unsupported SW Strategy type  : %s", type));
+        throw new IllegalArgumentException(String.format("Unsupported FW State type  : %s", type));
     }
 
-    public static LwM2MSoftwareUpdateStrategy fromStrategySwByCode(int code) {
-        for (LwM2MSoftwareUpdateStrategy to : LwM2MSoftwareUpdateStrategy.values()) {
+    public static UpdateStateFw fromStateFwByCode(int code) {
+        for (UpdateStateFw to : UpdateStateFw.values()) {
             if (to.code == code) {
                 return to;
             }
         }
-        throw new IllegalArgumentException(String.format("Unsupported SW Strategy code : %s", code));
+        throw new IllegalArgumentException(String.format("Unsupported FW State code : %s", code));
     }
-
 }
